- Updated date grouping to show full month name for Today and Yesterday, and use long month formatting in the date strings. This tightens and standardizes date labels across the Journal page.

X-Lovable-Edit-ID: edt-07545ff1-060d-432b-97be-233996109300

+++ src/pages/JournalPage.tsx
@@ -26,14 +26,14 @@ function getDateGroup(dateStr: string): string {
   const diffDays = Math.round((today.getTime() - entryDate.getTime()) / 86400000);
 
   if (diffDays <= 0) {
-    const day = d.toLocaleDateString("en-US", { day: "numeric", month: "short" });
+    const day = d.toLocaleDateString("en-US", { day: "numeric", month: "long" });
     return `Today, ${day}`;
   }
   if (diffDays === 1) {
-    const day = d.toLocaleDateString("en-US", { day: "numeric", month: "short" });
+    const day = d.toLocaleDateString("en-US", { day: "numeric", month: "long" });
     return `Yesterday, ${day}`;
   }
-  return d.toLocaleDateString("en-US", { weekday: "long", day: "numeric", month: "short" });
+  return d.toLocaleDateString("en-US", { weekday: "long", day: "numeric", month: "long" });
 }
 
 function groupEntries(entries: JournalEntry[]): [string, JournalEntry[]][] {
